+++ Servidor/controllers/casoRechazadoController.js
@@ -6,6 +6,7 @@ const usuarioModel = require('../models/usuarioModel')
 const uuidv4 = require('uuid/v4');
 const crypto = require('crypto');
 const path = require('path');
+const Permisos = require('../models/permisos');
 
 function getCasosRechazados(req, res) {
   if(req.query.token == "undefined" || !auth.autentificarAccion(req.query.token)){
@@ -39,6 +40,11 @@ function createCasoRechazado(req, res) {
   if(!auth.autentificarAccion(usuario.token)){
     res.status(500)
     res.send({ error: true , type: 1})
+    return
+  }
+  if(Permisos.LIST_CRUD.indexOf(usuario.tipo)<0){
+    res.status(100)
+    res.send({ error: true , type: 2})
     return
   }
 
@@ -124,6 +130,12 @@ function reactivateCasoRechazado(req, res) {
     res.send({ error: true , type: 1})
     return
   }
+
+  if(Permisos.LIST_CRUD.indexOf(usuario.tipo)<0){
+    res.status(100)
+    res.send({ error: true , type: 2})
+    return
+  }
   casoRechazadoModel.deleteOne({_id: new mongoose.Types.ObjectId(req.params.id)})
     .exec((err, caso) => {
       //Configura nota con nota anterior
@@ -169,6 +181,11 @@ function editCasoRechazado(req, res) {
   if(!auth.autentificarAccion(usuario.token)){
     res.status(500)
     res.send({ error: true , type: 1})
+    return
+  }
+  if(Permisos.LIST_CRUD.indexOf(usuario.tipo)<0){
+    res.status(100)
+    res.send({ error: true , type: 2})
     return
   }
   
@@ -253,6 +270,11 @@ function deleteCasoRechazado(req, res) {
   if(!auth.autentificarAccion(usuario.token)){
     res.status(500)
     res.send({ error: true , type: 1})
+    return
+  }
+  if(Permisos.LIST_CRUD.indexOf(usuario.tipo)<0){
+    res.status(100)
+    res.send({ error: true , type: 2})
     return
   }
